requirements.txt
future
six
configparser
Django>=2.2.10,<3.0
django-formtools>=1.0
pisa
Pillow==7.0.0
anyjson>=0.2.4
django-authority>=0.4
django-countries>=4.4
django-form-utils>=1.0.3
django-localflavor==1.4.1
dj-pagination>=2.3.0
django-picklefield>=0.1.6
django-simple-captcha==0.5.12
django-nocaptcha-recaptcha==0.0.20
django-tagging==0.4.6
django-haystack==2.8.1
django-debug-toolbar>=1.9.1
django-sql-explorer==1.1.3
feedparser>=4.1
pytz==2019.3
simplejson>=2.0.9
webcolors>=1.3.1
xlrd>=0.9.4
xlwt>=0.7.2
beautifulsoup4==4.8.2
oauth2>=1.5.167
#python_openid>=2.2  # Used by the social_auth app, but does not support Python 3
createsend==4.2.7
#django-celery==3.3.1
celery==4.4.0
mimeparse>=0.1.3
python-dateutil==2.8.1
pdfminer.six==20170720
stripe==2.42.0
pycryptodome==3.9.6
boto==2.49.0
django-timezone-field==2.1
django-ses==0.8.14
django-tastypie==0.14.3
docutils==0.16
chardet==3.0.4
psycopg2>=2.7,<2.8
gunicorn==20.0.4
gevent==1.4.0
django-storages==1.4.1
Whoosh==2.7.4
simple-salesforce==0.68.1
selenium==3.141.0
raven==6.0.0
tendenci-django-admin-bootstrapped>=3.0
django-bootstrap3==11.1.0
django-app-namespace-template-loader==0.4
django-annoying
unidecode
Markdown
bleach
unicodecsv
Embedly>=0.5.0
# dependencies for helpdesk
django-bootstrap-form>=3.1,<4
email-reply-parser
django-markdown-deux
xhtml2pdf==0.2.2
python-magic
XlsxWriter==0.9.6
nameparser

